Move server helpers to their own file; add type hints
