+++ upgrade.md
@@ -1,0 +1,5 @@
+# Upgrading to v4
+
+- Update custom drivers:
+    - Match the new type hints of the base ValetDriver
+    - Extend the new namespaced drivers instead of the old globally-namespaced drivers
